Combine into existing compat
