--- optionals/compat_rksl_pm_525/$PBOPREFIX$
@@ -1,1 +1,0 @@
-z\ace\addons\compat_rksl_pm_525
