--- optionals/compat_rksl_pm_525/config.cpp
@@ -1,17 +1,0 @@
-#include "script_component.hpp"
-
-class CfgPatches {
-    class ADDON {
-        name = COMPONENT_NAME;
-        units[] = {};
-        weapons[] = {};
-        requiredVersion = REQUIRED_VERSION;
-        requiredAddons[] = {"RKSL_PMII_525"};
-        author = ECSTRING(common,ACETeam);
-        authors[] = {"Dedmen"};
-        url = ECSTRING(main,URL);
-        VERSION_CONFIG;
-    };
-};
-
-#include "CfgWeapons.hpp"
