--- optionals/compat_rksl_pm_525/script_component.hpp
@@ -1,6 +1,0 @@
-#define COMPONENT compat_rksl_pm_525
-#define COMPONENT_BEAUTIFIED RKSL PMII 525 Compatibility
-
-#include "\z\ace\addons\main\script_mod.hpp"
-
-#include "\z\ace\addons\main\script_macros.hpp"
