+++ optionals/compat_rksl_pm_ii/CfgWeapons.hpp
@@ -34,4 +34,20 @@ class CfgWeapons {
             };
         };
     };
+
+    class RKSL_optic_PMII_525 : ItemCore {
+        ACE_ScopeHeightAboveRail = 4.2235;
+        ACE_ScopeAdjust_Vertical[] = { -4, 30 };
+        ACE_ScopeAdjust_Horizontal[] = { -6, 6 };
+        ACE_ScopeAdjust_VerticalIncrement = 0.1;
+        ACE_ScopeAdjust_HorizontalIncrement = 0.1;
+        class ItemInfo : InventoryOpticsItem_Base_F {
+            class OpticsModes {
+                class Snip {
+                    discreteDistance[] = { 100 };
+                    discreteDistanceInitIndex = 0;
+                };
+            };
+        };
+    };
 };
